requirements.txt
google-generativeai==0.8.5
python-dotenv==1.1.1
python-telegram-bot==22.4
pillow==11.3.0
json-repair==0.51.0